refactor: consolidate utilities into utils_core
